Pangu (#14)
* Remove debug logs

* Change the length of the field msgType from 4 bytes to 2 bytes

* ReceiveLoop has not yet been enabled when loading the journal msg

* update wal_test

+++ consensus/cbft/wal_decoder.go
@@ -126,7 +126,7 @@ type JournalMessagePrepareBlockHash struct {
 	Data      *MsgInfoPrepareBlockHash
 }
 
-func WALDecode(pack []byte, msgType uint32) (*MsgInfo, error) {
+func WALDecode(pack []byte, msgType uint16) (*MsgInfo, error) {
 	switch msgType {
 	case PrepareBlockMsg:
 		var j JournalMessagePrepareBlock
